+++ lib/dial.rb
@@ -2,3 +2,4 @@
 
 require_relative "dial/version"
 require_relative "dial/railtie"
+require_relative "dial/engine"