+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+- Add support for request profiling with vernier
+
 ## [0.1.1] - 2024-11-08
 
 - Set upper bound on dependencies